requirements.txt
Flask==2.3.3
requests==2.32.5
gspread==6.2.1
oauth2client==4.1.3
python-dotenv==1.0.1
